Move INodePacket, NodePacketType (and extensions) to Framework
- Move INodePacket to src/Framework/BackEnd/INodePacket.cs
- Move NodePacketType to src/Framework/BackEnd/NodePacketType.cs
- Move NodePacketTypeExtensions to src/Framework/BackEnd/NodePacketTypeExtensions.cs
- Delete src/Shared/INodePacket.cs
